requirements.txt
asgiref==3.3.4
dj-database-url==0.5.0
Django==3.2.3
django-crispy-forms==1.11.2
django-restframework==0.0.1
djangocms-installer==2.0.0
djangorestframework==3.11.0
numpy==1.21.0
pandas==1.2.5
Pillow==8.2.0
python-dateutil==2.8.1
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
tzlocal==2.1
